Se reestructura GLBViewer se separan componentes se deja todo funcionando
