+++ src/components/animation/AnimatedModel.jsx
@@ -1,0 +1,34 @@
+import { AnimationMixer } from "three";
+import { useGLTF } from "@react-three/drei";
+import { useEffect, useRef } from "react";
+import { useFrame } from "@react-three/fiber";
+
+const AnimatedModel = ({ url }) => {
+  const group = useRef();
+  const { scene, animations } = useGLTF(url);
+  const mixer = useRef();
+
+  useEffect(() => {
+    if (animations.length > 0) {
+      mixer.current = new AnimationMixer(scene);
+      animations.forEach((clip) => {
+        mixer.current.clipAction(clip).play();
+      });
+    }
+
+    scene.traverse((child) => {
+      if (child.isMesh) {
+        child.castShadow = true;
+        child.receiveShadow = true;
+      }
+    });
+  }, [animations, scene]);
+
+  useFrame((_, delta) => {
+    mixer.current?.update(delta);
+  });
+
+  return <primitive ref={group} object={scene} dispose={null} />;
+};
+
+export default AnimatedModel;
