+++ src/components/controls/Controls.jsx
@@ -1,0 +1,7 @@
+import { OrbitControls } from "@react-three/drei";
+
+const Controls = ({ controlsRef }) => {
+  return <OrbitControls ref={controlsRef} enableZoom={false} />;
+};
+
+export default Controls;
