+++ src/components/lights/Lights.jsx
@@ -1,0 +1,15 @@
+const Lights = () => {
+  return (
+    <>
+      <ambientLight intensity={0.4} />
+      <directionalLight
+        castShadow
+        position={[5, 10, 5]}
+        intensity={1.2}
+        shadow-mapSize-width={2048}
+        shadow-mapSize-height={2048}
+      />
+    </>
+  );
+};
+export default Lights;
